refactor: 정보 입력 페이지 파일 구조 정리
